+++ tests/common/test_nwbfile.py
@@ -1,0 +1,41 @@
+import os
+
+import pytest
+
+
+@pytest.fixture
+def common_nwbfile(common):
+    """Return a common NWBFile object."""
+    return common.common_nwbfile
+
+
+@pytest.fixture
+def lockfile(base_dir, teardown):
+    lockfile = base_dir / "temp.lock"
+    lockfile.touch()
+    os.environ["NWB_LOCK_FILE"] = str(lockfile)
+    yield lockfile
+    if teardown:
+        os.remove(lockfile)
+
+
+def test_get_file_name_error(common_nwbfile):
+    """Test that an error is raised when trying non-existent file."""
+    with pytest.raises(ValueError):
+        common_nwbfile.Nwbfile._get_file_name("non-existent-file.nwb")
+
+
+def test_add_to_lock(common_nwbfile, lockfile, mini_copy_name):
+    common_nwbfile.Nwbfile.add_to_lock(mini_copy_name)
+    with lockfile.open("r") as f:
+        assert mini_copy_name in f.read()
+
+    with pytest.raises(AssertionError):
+        common_nwbfile.Nwbfile.add_to_lock("non-existent-file.nwb")
+
+
+def test_nwbfile_cleanup(common_nwbfile):
+    before = len(common_nwbfile.Nwbfile.fetch())
+    common_nwbfile.Nwbfile.cleanup(delete_files=False)
+    after = len(common_nwbfile.Nwbfile.fetch())
+    assert before == after, "Nwbfile cleanup changed table entry count."